+++ src/main/java/cn/nukkit/level/format/anvil/Anvil.java
@@ -317,7 +317,7 @@ public class Anvil extends BaseLevelProvider {
 
     public static ChunkSection createChunkSection(int Y) {
         CompoundTag nbt = new CompoundTag();
-        nbt.putByte("Y", (byte) Y);
+        nbt.putByte("Y", Y);
         nbt.putByteArray("Blocks", new byte[4096]);
         nbt.putByteArray("Data", new byte[2048]);
         byte[] sl = new byte[2048];
